requirements.txt
fastapi==0.115.6
uvicorn==0.34.0

# database
beanie==1.29.0
email_validator==2.2.0

# hashing e JWT
passlib==1.7.4
PyJWT==2.10.1
python-multipart==0.0.20

# testing
pytest==8.3.4
httpx==0.28.1
mongomock==4.3.0
pytest-asyncio==0.25.2